tasks/9_create-an-intelligent-log/loganalyzer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loganalyzer"
version = "0.1.0"
description = "Intelligent log analysis CLI tool with anomaly detection and actionable insights"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "LogAnalyzer Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
]
keywords = ["logs", "analysis", "anomaly-detection", "cli", "monitoring"]

dependencies = [
    "click>=8.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "jinja2>=3.1.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "regex>=2023.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.scripts]
loganalyzer = "loganalyzer.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/loganalyzer"
Documentation = "https://github.com/yourusername/loganalyzer/blob/main/README.md"
Repository = "https://github.com/yourusername/loganalyzer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
